+++ packages/next-utils/src/measure.ts
@@ -1,6 +1,6 @@
-import { measure } from '@kirill.konshin/utils';
+import { createMeasurer } from '@kirill.konshin/utils';
 import { PHASE_PRODUCTION_BUILD, PHASE_PRODUCTION_SERVER } from 'next/constants';
 
-const isProd = process.env.NEXT_PHASE === PHASE_PRODUCTION_BUILD || process.env.NEXT_PHASE === PHASE_PRODUCTION_SERVER;
+export const isProd = process.env.NEXT_PHASE === PHASE_PRODUCTION_BUILD || process.env.NEXT_PHASE === PHASE_PRODUCTION_SERVER;
 
-export default measure.createMeasurer({ colors: isProd });
+export const measurer = createMeasurer({ colors: isProd });