+++ packages/electron-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @kirill-konshin/electron-utils
 
+## 0.0.3
+
+### Patch Changes
+
+- Various updates & fixes, Electron implementation
+
 ## 0.0.2
 
 ### Patch Changes