+++ packages/next-utils/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @kirill.konshin/next-utils
 
+## 0.0.5
+
+### Patch Changes
+
+- Various updates & fixes, Electron implementation
+- Updated dependencies
+    - @kirill.konshin/utils@0.0.6
+
+## 0.0.4
+
+### Patch Changes
+
+- Exports
+
+## 0.0.3
+
+### Patch Changes
+
+- Flat utils, error, misc fixes
+- Updated dependencies
+    - @kirill.konshin/utils@0.0.5
+
 ## 0.0.2
 
 ### Patch Changes