+++ packages/react-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @kirill.konshin/react-utils
 
+## 0.0.4
+
+### Patch Changes
+
+- Various updates & fixes, Electron implementation
+
+## 0.0.3
+
+### Patch Changes
+
+- Flat utils, error, misc fixes
+
 ## 0.0.2
 
 ### Patch Changes